+++ test/test_response_header.rb
@@ -84,15 +84,11 @@ class TestResponseHeader < Minitest::Test
 
   # The header must not contain a Status key.
   def test_status_key
-    skip 'implement later'
-
     assert_ignore_header("Status", "500")
   end
 
   # Special headers starting “rack.” are for communicating with the server, and must not be sent back to the client.
   def test_rack_key
-    skip 'implement later'
-
     assert_ignore_header("rack.command_to_server_only", "work")
   end
 
@@ -102,12 +98,10 @@ class TestResponseHeader < Minitest::Test
   # Header keys will be set through two ways: Regular and early hints.
 
   def test_illegal_character_in_key
-    skip 'implement later'
     assert_ignore_header("\"F\u0000o\u0025(@o}", "Boo")
   end
 
   def test_illegal_character_in_key_when_early_hints
-    skip 'implement later'
     assert_ignore_header("\"F\u0000o\u0025(@o}", "Boo", early_hints: true)
   end
 
@@ -115,23 +109,18 @@ class TestResponseHeader < Minitest::Test
   # Header values can be set through three ways: Regular, early hints and a special case for overriding content-length
 
   def test_illegal_character_in_value
-    skip 'implement later'
     assert_ignore_header("X-header", "First \000Lin\037e")
   end
 
   def test_illegal_character_in_value_when_early_hints
-    skip 'implement later'
     assert_ignore_header("X-header", "First \000Lin\037e", early_hints: true)
   end
 
   def test_illegal_character_in_value_when_override_content_length
-    skip 'implement later'
     assert_ignore_header("Content-Length", "\037")
   end
 
   def test_illegal_character_in_value_when_newline
-    skip 'implement later'
-
     server_run app: ->(env) { [200, {'X-header' => "First\000 line\nSecond Lin\037e"}, ["Hello"]] }
     data = send_http_and_read "GET / HTTP/1.0\r\n\r\n"